Colissions dont work, also do exercise
Create an obstacle course of platforms and terrain to be navigated, and 10 targets. Modify
the scene, the total time available, and time bonus given by each target to make it easy to
shoot the first 5 targets, but difficult to shoot the last 5 targets – aiming for only a few
seconds remaining of the total time on completion.